Refactor ColorTheme
Renamed ColorTheme to ThemeSelector to be used only for selecting theme and create new ColorTheme component as parent component (in future swapped with some Navigation component)

Refactor ColorTheme

Renamed ColorTheme to ThemeSelector to be used only for selecting theme and create new ColorTheme component as parent component (in future swapped with some Navigation component)
